Show better diffs for strings in tests
The ‘toMatchStringIgnoringWhitespace’ assertion is very helpful but
when there is an error the diffing isn’t as good as the native `toBe`
assertion.
This adds the diffing that jest uses for the ‘toBe’ assertion if the variable
is a string.
As jest-diff is part of jest I don’t see the need to declare it in our
package.json