Handle non-array requiredCharacteristics
When one requiredCharacteristic is selected, Express “helpfully”
flattens this down to a single string, rather than an array. This is
not in any way helpful, and has been the cause of many a flickering
integration test. To get round this, we need to cast the characteristic
into an array and flatten it before sending the characteristics on their
merry way to the API.