Add an additionalAdjustments field
In the designs, this featured a bulleted list, but I was concerned about
how this bulleted list would relate to the form label, and would also
look weird if there was one need, so I’ve changed it to a sentence.

I’ve also updated the needs copy in the previous question, as the
repetition of “needs” made it read a bit strangely in both pages.