Consolidate functions summary list functions
We had two similar functions for generating the summary list.
The error we are trying to fix here was caused by incorrectly assuming
they worked the same way.
We can combine the two, using the version in summaryListUtils.ts
and add a condition to check if the page is ‘attach-documents’ and if
it is treat it differently. This is the only bit of functionality that
was different between the functions.
In order to remove a dependency cycle we have to move over some extra
util functions to summaryListUtils.ts but they deal with summmary lists
so it make sense to me.
Other than this we do a bit of renaming to make things make sense