Allow summaryListItem for response to handle assessment links
In order to consolidate the two functions we need to be able to link
to assessments as well as applications.
summaryListItemForResponse isn’t tested directly, only its outputs in
getTaskResponsesAsSummaryListItems are so we will write a failing
test so we can check that it is working as expected in the next commmit.