Fix characteristics helpers
These weren’t working with real data, because we should be matching on
the `propertyName`, not the `name`. I’ve updated the factory, as well as
adding some tests, which were missing too. To make testing easier, I’ve
also updated `matchedCharacteristics` and `unmatchedCharacteristics` to
accept the characteristic pairs, not just the bed search result.