Allow shouldContainSummaryListItems to handle links in dt
Previously we assumed that the key would be text if the item.value
wasn’t text.
However the item.key could be ‘html’ which would mean cy.contains would
break as it cannot find html, only text. So now if the item.key is
html we lookup dt by the dd instead of looking up the dd by the dt