Add detail to the move on plans in place questions
This adds detail field to the ‘Yes’ and ‘No’ repsonses on the Move On
plans questions. To be a bit more precise, I’ve added two seperate
types, one for when the answer is no and one for when the answer is yes,
which means we can clear any irrelevant data from the response if the
user goes back and amends the response.