Update placement request data util
Now the criteria and the characteristics are mapped 1:1, and we’ve
rationalised the criteira we do need, we don’t have to do lookups
anymore, so we can just push criteria directly to the essential and
desirable criteria arrays. The AP type is also a characteristic, so
if the type is anything other than ‘normal’, we push the type of AP
to the `essentialCriteria` array.

Similarly, if `mentalHealthSupport` is selected, we add
`isSemiSpecialistMentalHealth` to the `essentialCriteria` array.

We also have a situation where if a user selects
`acceptsSexOffenders` or `acceptsChildSexOffenders` (which is
a premises-level characteristic), we also need to add
`isSuitedForSexOffenders` (which is a room-level characteristic)