Add a helper to translate/group criteria
This adds some types for each category of criteria, which are subsets of
the main `PlacementCriteria` type. We also add a `placementCriteria`
object, which contains a translation for each criteria item. As this is
a Record, we’ll get a type error if the API adds any new criteria that
are not translated.

We can then export the categoried options for use in the Assessment
form.