Update the integration tests
I’ve broken them down into three seperate tests to (hopefully) make it a
bit more readable. There was a lot going on, so it was hard to see what
was happening where. I’ve also hardcoded the essential and desirable
criteria in the search test to make the output a bit more predictable.