Add getOccupancy method to Premises service
I was unsure whether to use 'occupancy' or 'calendar' for the name here but I believe
we may use this endpoint in the future for occupancy and not just the calendar so
I opted for a more generic name than on the client which is named after the endpoint
it calls.
I also have added a mapping function here to save us mapping ISO 8601 dates to JS dates
in many more places. I am unsure of the amount of data we will be getting from the endpoint
so I think it's safer to do the conversion async.