Set `duration` to be undefined, not null
Before, we were casting the return of `this.lengthInDays()` as a string,
so if the return value was `null`, we’d set `this.body.duration` to
`’null’` (a literal string with the value `null`). Now we do the casting
in the return value and return `undefined` if we don’t have a duration.