207: add nunjucks utility for lostBeds
We would like users to be able to edit out of service/lost beds -
initially just to update their end date and notes. When a user is on the
list of all beds for a premises, they wil be able to click an action
menu item to manage out of service beds, they will then be shown a list
of all lost beds for a premises and be able to click to manage one.

This commit adds some utlitity functions that we will use shortly to
create a table of lost beds in a view.