207: add nunjucks views for lostBed flow
We would like users to be able to edit out of service/lost beds -
initially just to update their end date and notes. When a user is on the
list of all beds for a premises, they wil be able to click an action
menu item to manage out of service beds, they will then be shown a list
of all lost beds for a premises and be able to click to manage one.

This commit adds the views we will need to show a list of lost beds, a
single lost bed and the menu item to navigate to the list page.