207: controller methods to get and update lostBeds
We would like users to be able to edit out of service/lost beds -
initially just to update their end date and notes. When a user is on the
list of all beds for a premises, they wil be able to click an action
menu item to manage out of service beds, they will then be shown a list
of all lost beds for a premises and be able to click to manage one.

This commit adds the methods to the controller to get a list of lost
beds and update a lost bed, using the views created previously. The
`referer` from the header is used to redirect because the `show` view
for a `LostBed` could be accessed if there is a booking conflict [1].

[1] server/utils/bookingUtils.ts