Add UI types for BedOccupancyEntrys
The typegen isn't smart enough to work out these discriminated unions so we
have to manually type them and use these types instead. Not super happy with this but
not sure how to improve on it in the short term.
We also add the JS dates to these types as we will be using these going forward