Create seperate factories for UI and API types
Given we will be dealing with dates a lot with this calendar it makes sense
to have factories for creating objects that the UI can handle with converting
ISO8601 to JS Date objects. This means we have to create a bunch more slightly
different factories. Due to faker/TS rigidity we can't simply overwrite the string dates.