Remove upcomming arrivals and departures window
We thought that it may be useful to limit upcomming departures and arrivals
to the next 40 days. However AP managers need to be able to see all upcomming arrivals.
I can imagine this being adjusted in future so I think it's best to leave the filtering
mechanism in place and set the window to a very high number.