220: add a utility method for phase banner
We need to show a GDS phase banner[1] on all pages of the app in order
to capture feedback from users or receive emails from them. In the main
sub-sections of the app we will need specific links to feedback forms as
well as a link to email, but on the home page we just need copy to
direct users to email us.

This commit adds a utility function to return the correct copy based on
the current path, which will be passed in from a nunjucks template
(already defined as `currentUrl` in global vars)

[1]https://design-system.service.gov.uk/components/phase-banner/