220: add the phase banner to header partial
We need to show a GDS phase banner[1] on all pages of the app in order
to capture feedback from users or receive emails from them. In the main
sub-sections of the app we will need specific links to feedback forms as
well as a link to email, but on the home page we just need copy to
direct users to email us.

This commit adds the component to the header partial. We decided to sit
it below the `mojPrimaryNavigation` because it did not look quite right
sitting between the main header and the nav. Although the GDS guidance
says to put it directly under the header, there's some discussion about
this being the right place [2] and this seems like the right thing for
us.

[1] https://design-system.service.gov.uk/components/phase-banner/
[2] https://github.com/alphagov/govuk-design-system-backlog/issues/57