220: integration test phase banner
We need to show a GDS phase banner[1] on all pages of the app in order
to capture feedback from users or receive emails from them. In the main
sub-sections of the app we will need specific links to feedback forms as
well as a link to email, but on the home page we just need copy to
direct users to email us.

This commit adds the integration tests to check the banner appears with
the correct copy for the different navigation sections.

[1] https://design-system.service.gov.uk/components/phase-banner/