Fix startDates before the calendar starts
Before, if a schedule item had a startDate that came before the
start of the calendar, the item would be skipped as the start
date was not the same as the date in the cell. This adds a new
`scheduleForCalendar` function, which, as well as setting that
startDate to the calendar’s start date if it is a date before
the calendar’s date also adds a new `label` attribute, which
records the actual dates of the booking, so it can be recorded
accurately on the calendar.