Ensure desirable search filters keep their state between searches
Previously if a filter was selected then when the form was submitted it would
be unselected when the page was reloaded... despite showing the results filtered
as intended.
We now feed the selected characteristics (params.requiredCharacteristics) back into
the filters so they remain selected.
I am not entirely sure how the integration tests were passing before but I have
corrected them now so they take both required characteristics and optional
characteristics into account.§