Add util to map beds to select items
This will be used in the following commit.
As beds often have non-descriptive names (EG '1') I have opted to use
the room name too in order for staff to know which bed is which (there could
imaginably many beds named '1').
In future we could either return the room name with the bed from the API
or call `/premises/{premisesId}/beds/{bedId}` for each bed in the UI.