219: return `placementRequestDetail` from service
We would like to show past cancellations/appealed bookings on a
Placement Request (Match) so that the user can make an informed decision
when trying to match them again.

The API was updated to return a `placementRequestDetail` with
`cancellations` as an extra field. This commit updates the Service to
reflect that.