Add styling
This has a fair bit of SASS magic involved, as we need to
ensure the width of the calendar is consistent. We set the size
of the cells, and their padding, and then do some maths to
calculate the full width of the table. We also have a SASS loop
that sets the width of cells by their colspan - for example, if
the cell has a colspan of 4, then the width is the width of a normal
cell, plus the padding on each side, multiplied by 4:

```
($colWidth + ($colPadding * 2)
```

Everything else is fairly conventional. I’ve tried to use BEM classes
to be consistent with GDS style, make the CSS easier to grok and allow
us to reuse the existing table styles easily.