Update cell content and add tooltips
As we’re adding `overflow: ellipsis` to all cells, it doesn’t make
sense to reduce the content of the cells. We’re also adding some
tooltips so users can see the full cell content when they mouseover.

This also adds a bit of frotend JS to remove the tooltips when the text
does not overflow.