Send audit messages depending on redirect destination
By reading the "Location" header from the response, we can determine
where the request handler redirected the user to, and send audit
messages depending on this destination. Using the path pattern, we can
also determine the details to include in the audit message.

In many of our PUT and POST routes, we follow the pattern of
redirecting back to the source page in the event of an error, and to a
page for viewing the edited/created entity on success. We can use this
to send the appropriate audit event in each case.