Add forPagesInTask function for iterating through task pages
This differ from the version in TA as our first task starts with an
coniditional question, so we check that there is a response for the question
before dealing with it. If there isn't we skip to the next question if there
is one, if there isn't we set the pageName to an empty string.
We also filter out the check-your-answers screen because we never care
about the answer to that question