Allow additional metadata to be sent
As we reuse our controllers for the form actions, we can’t record
the task and page names at runtime, so we need to add an
`additionalMetadata` argument to the middleware, so we can register
multiple different unique middleware handlers for each route including
the page and task names. This may also be useful for other actions later
down the line, so I’ve kept the naming generic.