Update navLinks selector
We now no longer have `role=tab` on the tabitems, as this was
causing an automated accessibility check error.

We also update the script to use the `govuk-visually-hidden` class
rather than the `hidden` attribute, as Cypress was picking this up
as visible, so the integration tests were passing by mistake.