Allow CRN search to handle 500 errors for missing OASys and NOMS number
If we recieve a 500 error from the service in the controller we now
check to see if it has a 'data.detail' property with the correct content
and if so we add the error to the flash. We also send 'checkOasys' and 'checkNoms'
from the view as this method is also called from a different view when
booking someone into an AP where they might not be concerned if there is
an OASys or NOMS present and they may not have time to contact support.