Add RFAP Suitability page
This page asks a few questions about RFAP suitability. I am not 100% sure
why the 'unsuitabilityForRfapRationale' question isn't a conditional reveal
when the user answers 'no' to the 'rfapIdentifiedAsSuitable' but it could be in future.
If the application is short_notice/emergency then we show the user the application
timeliness page next, otherwise it's back to the dashboard.