Update release date in integration test data
As the release date was in the past when we were checking noticeTypeFromApplication
it was always returning 'emergency' as the 'differenceInDays' was less than 7.
This wasn't a problem until I corrected the logic that shows the Application
Timeliness page to show when the notice type is 'emergency'. Now we have
updated the release date we can use `isShortNoticeApplication` param in the
`completeSuitabilityOfAssessmentQuestion` method in Assess helpers as intended