Throw Error object after catching
When an uncaught API error occurs, we get
`Non-Error exception captured with keys: message, stack`
sent to Sentry. This is because we’re throwing a plain object,
rather than an error, and this makes it hard to see what’s going
on. Rather than throwing the sanitised error, we recreate an error
object and throw that instead, which should make reviewing the errors
easier.