Ensure referrer is retained when showing an error
If we’re redirecting from an error, instead of showing fetching the
referrer (which will be the original page), we want to get the value of
the backLink. Similarly, if we have no referrer (i.e. the user has
navigated directly to the page), we want to have a default value for the
backLink.