Fix Preferred AP persistence
During testing, I noticed that the preferred APs weren’t getting
persisted correctly, and were showing as empty at the Check Your
Answers stage. This seems to be because we were expecting them to
come back from the API as SelectOptions, not ApprovedPremises, so
when running the select in the initialize stage, nothing was matching
at all, so `body.selectedAps` was always an empty array.

This updates the initializer to fix this, as well as ensuring
body.selectedAps is always an array of `ApprovedPremises`. I’ve also
removed the `selectedAps` variable, as it didn’t seem to be doing
anything.