Skip e2e tests for now
Upstream issues with the OAsys API are preventing the E2E tests from
running. To unblock the pipeline, I’ve removed the requirement for
the `e2e_environment_test` step to pass before we go on to preprod.

This is a temporary fix, and should be restored as soon as the upstream
issue is sorted.