Add helper for categorised withdrawal reasons
This splits the `WithdrawalReason`s into record objects depending on
their category, allowing us a type-safe way of using all the valid
reasons. We can then return the `withdrawalRadioOptions` with relevant
dividers to show what category each options belong to. The function also
accepts a string for the `conditionalValue`, which will be a HTML string
value that we can pass in directly from the view layer.