Add CSS for a sortable table
At the moment, we use the sortable table component in MOJ Frontend. This
is client-side only and generates a button. As we want to make a
server-side request to the API, we need to generate a link. This borrows
the CSS from MOJ Frontend, replacing the button attribute with a link.