Add a helper to generate a sortable table header
This will be used in the view to generate a header with a link to sort.

If the table is not sorted by that attribute, it generates a link to
sort the table by that field, together with a `ariaSort` of `none`

If the table is sorted by that attribute, it generates a link to
sort the table by that field in descending order, together with a
`ariaSort` of `ascending`.

If the table is sorted by that attribute in descending order, it
generates a link to sort the table by that field in ascending order,
together with a `ariaSort` of `descending`.