Add middleware to manage access by user role
This adds another thin middleware layer to manager access by role to
give us an added layer of protection should a user access an endpoint
they’re not allowed to access.

If they are not allowed to access an endpoint, we render an error page.

I’ve also noticed there is an incorrect casing of the `authError`
endpoint (which is called when a user doesn’t have the right HMPPS Auth
role), so fixed this here too.