Change links based on Booking Status
Before, if a placement request had a booking that was cancelled, they
would still see the option to cancel it, even though it was already
cancelled. This fixes the issue by matching on the placement request
status, rather than the presence / absence of a booking. I’ve also
updated the view to hide the booking if there is a cancelled booking.