Add github action to update bank holidays
At midnight every night his action should curl the gov.uk bank holidays API
and check if the downloaded json is the same as the existing json.
If it is the same it does nothing.
If the files are different it opens a PR in this repo updating the
json.