Add function that calculates business days including bank holidays
This borrows a lot from how the differenceInBusinessDays in date-fns works
but with the added holidays argument.
The user can choose to pass the holidays in (for testing purposes) otherwise
they are pulled from the bank-holidays.json.

The `isWithinInterval` from date-fns only allows for intervals where the
'start' is before the 'end' date. However when using `differenceInBusinessDays`
we need to know if the difference is negative. So we have to determine which
is the earliest date before passing it to `isWithinInterval`.