Remove dependency on person.type for testing if person is restricted
The person.type property depends on API changes that may not be deployed
when this change is. We can safely determine if a person is an LAO by
seeing if they have a 'name' property which is backwards compatible.
Once the neccessary API changes have been deployed we should to revert to
checking for the person.type as it is much clearer.