Add utils for building a user management table
We need to render a list of users so lets create some utils to help us.
These will be used in a subsequent commit in the view.

The email cell simply plucks an email property off any object that has one
so it can be placed in the generic tableUtils file whereas the other utils
dig deeper into the User object so I think they should stay in the 'users/tableUtils'
file