Add utils for editing users roles and qualifications
We get the 'roles' for a user back in an array. Some of the roles in this array
are simple roles however some of them relate to how assessments are assigned.
We want to differentiate the two in the UI so we need to filter one from
the other.
We then need to display hints next to some of the checkboxes and some of the
checkboxes have labels that can't be created using simple casing functions
so I have created a 'dictionary' for the 'translating' the roles as the API knows them
to how the user needs to be able to see them.
In order to keep things DRY we will use the same function to create both the checkboxes
however in order to mock those function (roleCheckboxItem, filterAllocationRoles) it is easiest
to have them in a different file to the higher level functions (rolesToCheckboxItems, allocationRolesToCheckboxItems)
hence the need for the 'roleCheckboxes' folder.