Add search functionality to list page
Users would like to be able to search for other users by their name.
This neccessitates another controller method on the user management controller.
When a search is performed we render the template again with the results
from the API and the search term in the search bar.

The search bar in the style proposed in the mocks doesn't exist elsewhere.
In order to get the button inline with the text input I have had to put
an empty div above the button and add some margin to that as the
maximum amount of margin available in the GDS DS is 9 and that was not
quite sufficient. It could be argued that we should use some custom SCSS
here instead