Flatten checkbox inputs
When express receives a single response from a checkbox input it 'helpfully' sends
it as a string as opposed to the usual array. This means array methods can't be used
and causes errors (EG https://ministryofjustice.sentry.io/issues/4395174101/events/f1d83babeea64b809e5b320b1884e2bb/?project=4503931667742720)
on the Check Your Answers screen.

Because flattenCheckboxItems is in formUtils.ts which is mocked in a lot of tests we
also have to update the tests to mock flattenCheckboxItems too but I have tried to do
so in a manner that doesn't get in the way of the tests themselves