Make preexisting 'lost beds'-specific report code more generic
This was written when the only reports available were ones on lost beds.
We need to make the naming more specific in a number of places. We need
to be less specific on the 'Download lost beds data' button in the view as
soon users will be able to download other data.

I opted to remove the 'show' object in the 'reports' object in the API paths
as this will have to be duplicated for each type of report.