Add application reports functionality
This adds the functionality for users to download Applications reports.
I opted to do the logic of selecting the correct client call in the Service
but we could have a service method for each client method and move this logic
to the controller. However in the following commit we will add validation which
increases the size of the controller method so rather than enlarge it further I think
the logic can sit happily in the service.