Add screen for searching for new user
Users need to be able to add other users from Delius.
This commit allows that to happen, given that they enter the correct user name on first attempt.
We add a new controller, partly because having two 'search's in one controller could be
confusing and partly because the 'userManagement' controller would be too big with the three
new endpoints we will have when finish adding this functionality.
We also add the confirmation screen that the user will be redirected to if the search was
successful. In the next commit we will handle the case when the user is not found