Add DELETE method to the rest client and audit middleware
We haven't needed to make a DELETE request before so we need to do a bit of setup.
I have opted to not add a lot of the config we have in the rest client for other verbs
because we don't seem to use it very often. If we do find we need it for DELETEs we can
copy and paste it in from the other verbs.