Add delete functionality
The user can click 'Remove access' from the Permissions/Edit page and they will
be asked to confirm if they want to remove access for that user.
If they confirm the request will be made to the API and they will be redirected to the
index page where they will be shown a confirmation message.

We need to add new client, service and controller methods for this as well as page objects
for the tests so this is quite a chunky commit