Handle null sections in OAsys
We’ve seen an error in prod where for some reason the `otherNeeds`
section is returning `null`, meaning the Optional Oasys Sections page
returns a 500. In the interests of moving this on, and allowing the user
to continue with the application, we should filter out any of these
`null` sections.