Ensure the response method returns a string for each question
Previously this returned an array which was allowed in the PageResponse type but it
fell apart when it came to generating the response in embeddedSummaryListItem because
embeddedSummaryListItem assumed it was dealing with an object.
As these recently changed `response` methods were the only places returning an array I decided
it is best to remove the Array<string> type from the PageResponse union.