Ensure correct hrefPrefix gets passed
This adds a new `getPaginationDetails` to fetch pagination details from
a request, and create a `hrefPrefix` with the appropriate suffix (`?` if
there are no query strings, or `&` if we need to add more query string
items to an existing query.

This will allow us to keep the sorting information around when paging
through sorted results.

I’ve also updated the `sortHeader` helper to make sure existing params
aren’t repeated.