Add function to addBusinessDays including bank holiday
Similar to our custom 'differenceInBusinessDays' function we need to be able
to add business days to a date that respects British bank holidays (BH).

We take the business days not including bank holiday using the date-fns function,
calculate the number of BHs that fall in the time period and then add that number
to the previously calculated business days.