Split 'roles' column into 'roles' and 'allocations'
In the user management table we show all roles in one column.
If a user has a lot of roles this can get quite lengthy so we have decided it's best
to show 'roles' and 'allocations' seperately, which mirrors how we display them on the 'errors'
page.
We also add 'qualifications' to the 'allocations' column.