Mock response from DaysUntilDue
Now that we're using business days in the daysUntilDue function we can no
longer predict how many days an assessment will have until it is due as there
is an unknown number of working days between the current date and the due date.
So, for test purposes, we have to mock the response from daysUntilDue.
To make this easier we'll put it in its own file.

We also add assertions that each of the functions contained within daysUntilDue
is called with the arguments we were expecting. It was hard to assert that
differenceInBusinessDays on line 8 is called with the correct arguments because
the 'new Date' called in the test file will be a few ms before the 'new Date' in the
code under test. In order to combat this I have added a todaysDate parameter that
can be passed in for test purposes and defaults to 'new Date' when not passed in. This
allows us to add the assertion whilst retaining the function's original behaviour.