Add adjacentPageFromSentenceType util
We need this logic in two places.
From the sentenceType page it tells us what the next page is.
For each branch in the journey there is one page and then most (except nonStatutory, which
we'll deal with later) pages continue to the releaseDate page. Therefore we can use the same
logic to tell us which page the user was on before they visited to releaseDate page in the
'previous' method.

For ease of mocking I have adjacentPageFromSentenceType in its own file and to avoid
creating a dependency cycle I have to add the sentenceTypes and SentenceTypesT type to this file.

Because we now calculate the previous page from releaseDate we no longer need the class contructor's
third argument 'previousPage' so we can remove that argument and update the tests.