Remove previousPage argument from forms
Initially we relied on this for 'back' button functionality when it was hard to calculate
which page the user had come from. However it is unreliable as it simply tells us the previous
page the user was on before the current one which may not neccessarily be the previous page in the
form (or even a page in the service).
We gradually retired its usages to the point where it was only used in one page, having removed
this usage in the previous commit we can now remove all the associated code.