Update non-statutory journey from sentence type page
We have previously had a high number of applications where the sentence type was 'Non statutory'.
This should be a very rare occurence and after post-hoc inspection each of these cases was not a valid
non-statutory case.
In order to better guide the user to the correct sentence type we add some guidance about what is a valid
non-statutory case and if non-statutory is selected we take the user to the 'managed by MAPPA' page which
we will add in the next commit.