Add 'Confirm your details' page
We add a page that asks users to confirm their details are correct and update
them if now.
We also ask them if they are they have case management responsibility and the next
page is dependent on their answer. If they are they continue with the application
if not then we have to ask them details of the person with case management responsibility.
Note the 'detailsToUpdate[]' syntax: this ensures that the detailsToUpdate checkbox input is always an array (even when it only has one entry). This negates the need to manually flatten it