Allow cypress to target checkbox with '[]' suffix
When Express parses inputs it strips the '[]' from the end which is nice and helpful.
However this means we can't use the '[]' in the applicationData to target the checkboxes which means
we can't use them in the names of the inputs that Cypress is targetting. Therefore cypress can't find them.
My solution here is to add an option to the checkCheckboxesFromPageBody method that, when
true, will add the '[]' to the input name.
In an ideal world all our checkboxes would have this suffix so this option wouldn't be needed
but for not I don't think it's too burdensome to have it.