Use Delius details if the user doesnt update fields
Previously if the user said they didn't want to update any of the details then
we didn't pull through the email, name and phoneNumber from Delius for the response value.
Now we do.
We have to map from the Delius names for fields to ones we can easily lowercase and still have them
make sense.