Add paragraphs to text with linebreaks
On some questions, particularly Oasys responses, text that has line
breaks is played back to users as a single block of text, making it
really hard to read. The adds a `linebreaksToParagraphs` helper which
we use in the `taskResponsesAsSummaryListItems` helper to replace single
linebreaks with `<br/>` HTML elements and double linebreaks with new
`<p>` paragraph elements.