Tasklist link returns the first completed section
Currently the `taskLink` helper returns the first question in the task.
This is causing users some confusion, as the first question in the
Basic Information task is telling the user that the Tier is not
applicable, which is only shown when the user has an ineligible tier.

To get around this, rather than blindly returning the first question in
the task, and the user has started the task, we return the first
question they have completed within that task.

This is not ideal, and we’ll need to iterate on this, for example if
we add a mandatory question to the start of a section, but this will
at least resolve some of the confusion.