Filter 'RequiredActions' task from the 'Assess Application' section
If users have told us that they haven't received the requisite information from the PP and that
they won't be accepting the application it doesn't make sense for them to give us information about the
placement requirements so we should remove this task from the section.

We havent done an 'optional task' before. I thought the best place for it was in the getSections function
used in the Tasklist Service. This means the logic is only in one place (we don't have to duplicate logic
into the views) and if we wanted to filter other tasks in the future we can do this easily.
The downsides of having it here are we have to call the function for every Section of both Applications
and Assessments. In order to avoid too much a performance hit we first check if a) we are dealing with an Application
and b) if we are dealing with the 'Assess Application' section of an Assessment. If not we return the section
untouched, if we are then we perform a further check to see if the information was received (or not) before
 filtering out the uncessary task.

The name for this function (filterSectionTasks) isnt as descriptive as it could be but I think it's the best
we can do without going too lengthy.