Add timeline to application view
This adds everything needed to display the timeline: the client, service
and controller methods as well as the view components.

I have opted to use a query param to enable switching between the two
views (readonly, timeline). If the user selects 'timeline' then we add a query
param and show the notes macro, otherwise we show the 'application' macro.

We are unable to use the MoJ Frontend Timeline component out of the box
(as a Nunjucks component) as it has a hidden dependency on some MoJ
Nunjucks filters which are dependent on Moment.js. As Moment.js is
EoL, has a big footprint and we already have a date manipulation library
(date-fns) it doesn't make sense to add it. It is also quite map the data
into the required shape (as we do in application/utils.ts) and then
iterate over the array to display the timeline (in show.njk).