Remove status check in informationSetAsNotReceived
We want this function to do what it says on the tin.
Previously it checked if the response to the 'informationReceived' question was 'no' AND if the
assessment status was 'awaiting_response'.
informationSetAsNotReceived is used in two places: a) in the assessmentsController.show[1] method and, b)
in the filterSections function.
In a) we want to check the assessment status is 'awaiting_response' but we do this already in the controller
method so we don't need to do it again in the util.
In b) we don't care about the assessment status and checking that it is 'awaiting_response' before checking
the answer to the question means that if a user changes their mind about the whether they have received
sufficient information they won't be able to complete the 'required-actions' task.

[1]https://github.com/ministryofjustice/hmpps-approved-premises-ui/blob/ec240ec2f68893f0739dd420b8a3c6feff8dd07e/server/controllers/assess/assessmentsController.ts#L31