Prefix LAOs in the Assessments table with 'LAO: '
Currently if someone is an LAO but the user has permissions to view LAO's they have no way of knowing.
If the user has permissions to view the person then they will come back from the API as
a FullPerson with the isRestricted flag set to 'true'. If this is the case we add 'LAO: ' when showing their name