Allow PlacementDate to handle cases with a past release date
Previously only cases that had a release date in the future would see this page.
Now that we handle cases with a release in the past it doesn't make sense
to ask the user if they would like the placement to start on the release date in all cases.
Now we check to see if a release date exists and if it is in the future, if it is then
we show the existing page. However if the release date doesnt exist or is in the past
then we ask the user what date they would like the placement to start.