Remove validation on release date
Previously we didn't allow applications to be made for a person whose release date
was in the past. However we have recently learned we need to allow for this as someone who is
'in the community' may need to stay in an AP. There is no way to determine if someone is in the
community from their sentence type or other information.
In order to make the change as simple as possible and to avoid creating design work we
will simply remove the validation on the release date.