Check release date in relation to the created date
Because we run validations when showing the application, the release
date will be in the past for applications where the release date has
already passed, so we get inconsistent behaviour. We fix this by
checking the release date in relation to when the application was
created. There are still some edge cases where the user may go back and
edit the repsonse, but this at least fixes the error in prod that we’re
seeing.