Add new tab to Application view for Placement Applications
We need a new tab where users can view Placement Applications.
We will later list PAs on this page and allow them to be withdrawn but
for now lets just add the 'Create Placement Request' button.
Because we've moved the button from the page to a seperate tab we need
to update the 'Placement Applications' tests.