Map Placement Applications for SummaryListComponent
We add some utils to help us marshall the data into the neccessary format.
It is worth noting the change in the durationAndArrivalDateFromPlacementApplication
function. If someone creates a new Placement Application and then returns the
application/show/request-a-placement view without completing it the view will be broken completely
as we will be looking to retrieve a 'reason' that doesn't exist. So we need this function
to return an empty 'expectedArrival' and 'duration' so we can handle the empty values
downstream instead.